// src/interfaces/professionals.ts
export interface ProfessionalsAssociatedsProps {
  id?: number;
  name?: string;
  nickname?: string;
  cpf?: string;
  cellphone?: string;
  city?: string;
  state?: string;
  function?: string;
  localization: string;
  accreditation_status?: React.ReactNode;
}
export interface ProfessionalsAssociatedsTableProps {
  id?: number;
  name?: string;
  nickname?: string;
  cpf?: string;
  cellphone?: string;
  city?: string;
  state?: string;
  function?: string;
  localization: string;
  accreditation_status?: string;
}